// src/app/api/model/user-dto-in.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AccountStatus } from './account-status';
import { UserType } from './user-type';
import { AddressDtoIn } from './address-dto-in';


export interface UserDtoIn { 
    /**
     * Account status. Valid transitions: INACTIVE -> IN_VALIDATION, TO_BE_DELETED, BANNED; IN_VALIDATION -> ACTIVE, INACTIVE, TO_BE_DELETED, BANNED; ACTIVE -> INACTIVE, TO_BE_DELETED, BANNED; BANNED -> ACTIVE, TO_BE_DELETED; TO_BE_DELETED -> DELETED, ACTIVE, IN_VALIDATION; DELETED -> (terminal, no transitions)
     */
    accountStatus: AccountStatus;
    addresses?: Array<AddressDtoIn>;
    addressesIds?: Array<number>;
    email?: string;
    firebaseUserId?: string;
    firstName?: string;
    id?: number;
    lastName?: string;
    name?: string;
    noteFromAdmin?: string;
    phoneNumber?: string;
    profilePicture?: string;
    userType: UserType;
}